fix x64 build

+++ library/modules/runtime/graphics/source/win_d3d11/shader_d3d11.cpp
@@ -98,9 +98,9 @@ namespace tiki
 		TGInputLayout* pLayout = nullptr;
 		const HRESULT result = pDevice->CreateInputLayout(
 			pElements,
-			elementCount,
+			(UINT)elementCount,
 			shaderData.pShaderCode,
-			shaderData.shaderCodeLength,
+			(UINT)shaderData.shaderCodeLength,
 			&pLayout
 		);
